+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- Fix high percentage of slow frames (#1915)
+
 ## 7.18.0
 
 ### Features